+++ packages/magento-cart-coupon/RemoveCouponForm/RemoveCouponForm.tsx
@@ -13,7 +13,8 @@ const useStyles = makeStyles((theme: Theme) => ({
     color: theme.palette.secondary.main,
     textTransform: 'uppercase',
     borderRadius: 4,
-    ...theme.typography.body1,
+    ...theme.typography.body2,
+    fontWeight: 600,
     display: 'flex',
     alignItems: 'center',
     '& .MuiIconButton-root': {